+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v1.9.0
+
+- Much faster rendering
+- Add support for http-tiles data in track config
+- Faster track resizing via y axis scaling
+
 v1.8.1
 
 - Make expected local tile format consistent with higlass's